chore: Cleaning up Auth integration tests (#239)
* chore: Cleaning up Auth integration tests

* fix: Added newline at the eof
